+++ Services/IEmployeeService.cs
@@ -5,7 +5,7 @@ namespace EmployeeApi.Services;
 public interface IEmployeeService
 {
     Task<IEnumerable<EmployeeDto>> GetAllAsync(string? search = null);
-    Task<EmployeeDto?> GetOneAsync(int id);
+    Task<EmployeeDto?> GetOneAsync(long id);
     Task<EmployeeDto> CreateAsync(CreateEmployeeDto input);
-    Task<bool> DeleteAsync(int id);
+    Task<bool> DeleteAsync(long id);
 }
